* Made Packet.TickCount internal * Added Packet.ResendCount so we can track the number of times a packet is resent (not implemented yet)
git-svn-id: http://libopenmetaverse.googlecode.com/svn/trunk@1619 52acb1d6-8a22-11de-b505-999d5b087335

+++ libsecondlife/ObjectManager.cs
@@ -703,8 +703,8 @@ namespace libsecondlife
             packet.ObjectData.PathRevolutions = LLObject.PackPathRevolutions(prim.PathRevolutions);
             packet.ObjectData.PathScaleX = LLObject.PackPathScale(prim.PathScaleX);
             packet.ObjectData.PathScaleY = LLObject.PackPathScale(prim.PathScaleY);
-            packet.ObjectData.PathShearX = LLObject.PackPathShear(prim.PathShearX);
-            packet.ObjectData.PathShearY = LLObject.PackPathShear(prim.PathShearY);
+            packet.ObjectData.PathShearX = (byte)LLObject.PackPathShear(prim.PathShearX);
+            packet.ObjectData.PathShearY = (byte)LLObject.PackPathShear(prim.PathShearY);
             packet.ObjectData.PathSkew = LLObject.PackPathTwist(prim.PathSkew);
             packet.ObjectData.PathTaperX = LLObject.PackPathTaper(prim.PathTaperX);
             packet.ObjectData.PathTaperY = LLObject.PackPathTaper(prim.PathTaperY);
